chore: clean up schemas
