+++ frontend/package.json
@@ -106,6 +106,7 @@
         "eslint": "^8.57.1",
         "eslint-config-next": "^15.1.5",
         "postcss": "^8.x",
+        "prettier-plugin-organize-imports": "^4.1.0",
         "tailwindcss": "^3.x",
         "typescript": "^5.7.2"
     },
