+++ frontend/src/components/canvas/FlowCanvas.tsx
@@ -14,7 +14,8 @@ import {
 import '@xyflow/react/dist/style.css'
 import { useSelector, useDispatch } from 'react-redux'
 import Operator from './footer/Operator'
-import { setSelectedNode, deleteNode, setNodes, FlowWorkflowNode, FlowWorkflowEdge } from '../../store/flowSlice'
+import { setSelectedNode, deleteNode, setNodes } from '../../store/flowSlice'
+import { FlowWorkflowNode, FlowWorkflowEdge } from '@/types/api_types/nodeTypeSchemas'
 import NodeSidebar from '../nodes/nodeSidebar/NodeSidebar'
 import { Dropdown, DropdownMenu, DropdownSection, DropdownItem, DropdownTrigger } from '@heroui/react'
 import { useKeyboardShortcuts } from '../../hooks/useKeyboardShortcuts'
