+++ frontend/src/components/nodes/BaseNode.tsx
@@ -8,7 +8,7 @@ import usePartialRun from '@/hooks/usePartialRun'
 import { TaskStatus } from '@/types/api_types/taskSchemas'
 import { AlertState } from '@/types/alert'
 import isEqual from 'lodash/isEqual'
-import { FlowWorkflowNode } from '@/store/flowSlice'
+import { FlowWorkflowNode } from '@/types/api_types/nodeTypeSchemas'
 import { getNodeTitle, duplicateNode, deleteNode } from '@/utils/flowUtils'
 import { RootState } from '@/store/store'
 import store from '@/store/store'
