+++ frontend/src/components/nodes/DynamicNode.tsx
@@ -12,7 +12,7 @@ import { useSelector } from 'react-redux'
 import BaseNode from './BaseNode'
 import styles from './DynamicNode.module.css'
 import { CardBody, Input } from '@heroui/react'
-import { FlowWorkflowNode } from '../../store/flowSlice'
+import { FlowWorkflowNode } from '@/types/api_types/nodeTypeSchemas'
 import { selectPropertyMetadata } from '../../store/nodeTypesSlice'
 import { RootState } from '../../store/store'
 import NodeOutputDisplay from './NodeOutputDisplay'
