+++ frontend/src/store/store.ts
@@ -2,7 +2,7 @@ import { combineReducers, configureStore } from '@reduxjs/toolkit'
 import { persistStore, persistReducer } from 'redux-persist'
 import storage from 'redux-persist/lib/storage'
 import flowReducer from './flowSlice'
-import { FlowState } from './flowSlice'
+import { FlowState } from '@/types/api_types/flowStateSchema'
 import nodeTypesReducer, { NodeTypesState } from './nodeTypesSlice'
 import userPreferencesReducer from './userPreferencesSlice'
 import panelReducer from './panelSlice'
